[#175] refactor: 디바이스 도메인 패키지 구조 개선
