+++ emulator/src/main/java/com/where_car/emulator/device/domain/car/CarIdentity.java
@@ -1,9 +1,10 @@
-package com.where_car.emulator.device.domain.common;
+package com.where_car.emulator.device.domain.car;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
 import lombok.Getter;
 import lombok.Setter;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
 
 /**
  * <pre>
